refactor: Refactor tests to extract mock data loading to utils module
Move the common mock data loading logic to a utils module
to avoid duplication across multiple test modules.
